기능: 구현할 기능 목록(README.md) 수정에 따라 기능 재구현. 실질적으로는 객체 이름을 수정.
