+++ src/main/java/Baseball/NumberHitter.java
@@ -1,50 +1,50 @@
 package Baseball;
 
-import Baseball.errors.InvalidInputException;
-
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Scanner;
+import java.util.Random;
 
 public class NumberHitter implements Hitter {
+    //todo: move to config
+    private final int ROUND = 3;
+    private final int MIN_NUMBER = 1;
+    private final int MAX_NUMBER = 9;
 
-    private Scanner scanner;
 
-    public NumberHitter(Scanner scanner) {
-        this.scanner = scanner;
-    }
 
     @Override
     public List<Integer> hit() {
-        System.out.print("숫자를 입력주세요 : ");;
-        String input = scanner.next();
-        if (!isValid(input)) {
-            throw new InvalidInputException("입력값이 옳바르지 않습니다.");
-        }
 
-        return translateToBalls(input);
-    }
-    private boolean isValid(String input) {
-        // todo: change to config
-        if (input.length() != 3) {
-            return false;
-        }
-
-        for (char syllable: input.toCharArray()) {
-            if (!Character.isDigit(syllable)) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    private List<Integer> translateToBalls(String input) {
         List<Integer> balls = new ArrayList<>();
-        for (char syllable: input.toCharArray()) {
-            int ball = Character.getNumericValue(syllable);
+        for (int i = 0; i < ROUND; i++) {
+            int ball = confirmBall(balls);
             balls.add(ball);
         }
         return balls;
+    }
+
+    private int confirmBall(List<Integer> balls) {
+        Random randomGenerator = new Random();
+
+        //todo: move to config
+        int ball = MIN_NUMBER - 1;
+        while (!isConfirmed(balls, ball)) {
+            //todo: check if it is good
+            ball = randomGenerator.ints(MIN_NUMBER, MAX_NUMBER + 1).limit(1).findFirst().getAsInt();
+        }
+
+        return ball;
+    }
+
+    private boolean isConfirmed(List<Integer> balls, int ball) {
+        if (isOutOfRange(ball)) {
+            return false;
+        }
+
+        return !balls.contains(ball);
+    }
+
+    private boolean isOutOfRange(int ball) {
+        return ball < MIN_NUMBER || MAX_NUMBER < ball;
     }
 }
